Refactored structure view changes (LuaFileElement)
* Extracted the visitor (LuaStructureVisitor)
* Extracted the new Element classes (ClassTreeElement, LuaNameDefElement, ShortMethodElement)
